reimplement using new AddUIToOptionsMenu mod
